Bumps version
Technically, this is only upgrading rspec and pinning it in the gemspec file.
This isn't a bug in terms of functionality but if I were to download the gem and
have the tests fail when checking out the gem, I would not be impressed. This
will make sure that people who install the gem can run the tests and have them
pass.